+++ src/com/nutrons/nu17/commands/GearControl/LowerGearPlacerCmd.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.commands;
+package com.nutrons.nu17.commands.GearControl;
 
 import com.nutrons.nu17.Robot;
 
@@ -7,11 +7,9 @@ import edu.wpi.first.wpilibj.command.Command;
 /**
  * 
  */
-public class RaiseGearPlacerCmd extends Command {
+public class LowerGearPlacerCmd extends Command {
 
-	private final double PLACER_MAX_POSITION = 1;
-
-    public RaiseGearPlacerCmd() {
+    public LowerGearPlacerCmd() {
     	requires(Robot.GP);
     }
 
@@ -19,15 +17,15 @@ public class RaiseGearPlacerCmd extends Command {
      * Lowers the gear placer.
      */
     protected void initialize() {
-    	Robot.GP.set(this.PLACER_MAX_POSITION);
+    	Robot.GP.set(0);
     }
 
     protected void execute() {
     	//empty
     }
-    // Finishes when the placer is at the highest position
+    // Finished when the placer is at the lowest position
     protected boolean isFinished() {
-        return Robot.GP.getPosition() == this.PLACER_MAX_POSITION;
+        return Robot.GP.getPosition() == 0;
     }
 
     protected void end() {
