+++ src/com/nutrons/nu17/commands/Intake/GroundIntakeSpitCmd.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.commands;
+package com.nutrons.nu17.commands.Intake;
 
 import com.nutrons.nu17.Robot;
 import edu.wpi.first.wpilibj.command.Command;
