+++ src/com/nutrons/nu17/subsystems/Climber/Climber.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.subsystems;
+package com.nutrons.nu17.subsystems.Climber;
 
 import com.nutrons.nu17.RobotMap;
 import edu.wpi.first.wpilibj.DigitalInput;
