+++ src/com/nutrons/nu17/subsystems/Drivetrain/Drivetrain.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.subsystems;
+package com.nutrons.nu17.subsystems.Drivetrain;
 
 import com.nutrons.nu17.OI;
 import com.nutrons.nu17.RobotMap;
