+++ src/com/nutrons/nu17/subsystems/Intake/GroundIntake.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.subsystems;
+package com.nutrons.nu17.subsystems.Intake;
 
 import com.nutrons.nu17.RobotMap;
