+++ src/com/nutrons/nu17/subsystems/Shooters/Shooter.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.subsystems;
+package com.nutrons.nu17.subsystems.Shooters;
 
 import com.nutrons.nu17.Robot;
 import com.nutrons.nu17.RobotMap;
