+++ src/com/nutrons/nu17/subsystems/Shooters/TwinShooter.java
@@ -1,4 +1,4 @@
-package com.nutrons.nu17.subsystems;
+package com.nutrons.nu17.subsystems.Shooters;
 
 import com.nutrons.nu17.RobotMap;
